Fix bug where # cards at showdown < 5
runSelection() was continually selecting raise which led to an all-in.
Code aliases raise to call when this happens. If statement condition
for checking all-in was inconsistent between raise and call; forgot
    the potInvestment term in raise(). When raise is aliased to call,
    conditionalDeal was not called. This means the stage advanced, but
    the cards were not dealt. Fixed by adding conditionalDeal in
    runSelection for raiseChild expansion. Finally, the
    conditionalDeal in all-in handling for runSimulation was
    short-circuited by the "all-in dealing". The stage would advance
    and all-in dealing would start from stage + 1, missing a stage of
    dealing. Debugged using debug stop2 and "conditional breakpoints".
    We found the missing card bug to always appear when selection's
    repeated raises led to all in. We then examined what steps led to
    this all in by using thread backtrace.

added debug lines to deal with the chip total, removed old debug lines
problem comes with calculating the initial number of chips to backprop

Add allocateChips in runSelection for showdown base case

We realized that while runSimulation() ran showdown and allocated
chips, when subsequently visiting the nodes through runSelection,
allocateChips would not be called. This was fixed by adding a showdown
and allocate in the showdown base case for runSelection. However, now
the problem appears to be a double allocate whose origins remain
mysterious. We thought it was because perhaps in runSimulation,
allocateChips was already called, meaning that when we visited the
node through runSelection and called allocateChips again, we'd be
double allocating. This seems not to be the case since in
runSimulation we reset the chips values to what they were before
the allocate after calling backprop.

Potential fix for double allocate; still bugged

Revert "Add allocateChips in runSelection for showdown base case"

This reverts commit 58120e08c50c49d0c9ce596d04178ae643ab7af6.

Update README.md