Add allocateChips in runSelection for showdown base case
We realized that while runSimulation() ran showdown and allocated
chips, when subsequently visiting the nodes through runSelection,
allocateChips would not be called. This was fixed by adding a showdown
and allocate in the showdown base case for runSelection. However, now
the problem appears to be a double allocate whose origins remain
mysterious. We thought it was because perhaps in runSimulation,
allocateChips was already called, meaning that when we visited the
node through runSelection and called allocateChips again, we'd be
double allocating. This seems not to be the case since in
runSimulation we reset the chips values to what they were before
the allocate after calling backprop.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch debugging
# Your branch is up-to-date with 'origin/debugging'.
#
# Changes to be committed:
#	modified:   Nodes/Node.cpp
#
# Changes not staged for commit:
#	modified:   Config.cpp
#
# Untracked files:
#	10utput.txt
#	output11.txt
#	output12.txt
#	output13.txt
#	output8.txt
#	output9.txt
#