Fix bug where cards wouldn't be dealt when selecting to all-in node
When running select, we have to push new cards onto the child node in
order to have different simulations. However, when pushing cards to
an all-in node, we didn't have all-in handling, so the all-in node had
too many cards. Now there is a problem with some sort of double
allocate somewhere; the chip amounts are too high